pkg/authorization: move to authorization/
