+++ integration/plugin/authz/authz_plugin_test.go
@@ -22,10 +22,10 @@ import (
 	"github.com/docker/docker/api/types/container"
 	eventtypes "github.com/docker/docker/api/types/events"
 	networktypes "github.com/docker/docker/api/types/network"
+	"github.com/docker/docker/authorization"
 	"github.com/docker/docker/client"
 	"github.com/docker/docker/integration/util/request"
 	"github.com/docker/docker/internal/test/environment"
-	"github.com/docker/docker/pkg/authorization"
 	"github.com/gotestyourself/gotestyourself/skip"
 	"github.com/stretchr/testify/require"
 )
