+++ integration/plugin/authz/main_test.go
@@ -12,9 +12,9 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/docker/docker/authorization"
 	"github.com/docker/docker/integration-cli/daemon"
 	"github.com/docker/docker/internal/test/environment"
-	"github.com/docker/docker/pkg/authorization"
 	"github.com/docker/docker/pkg/plugins"
 )
